CLI script for the typescript-reference-doc-generator
