--- packages/typescript-reference-doc-generator/bin/api-documenter
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-require('../lib/start.js')
