+++ packages/typescript-reference-doc-generator/package.json
@@ -1,37 +1,40 @@
 {
-  "name": "@microsoft/api-documenter",
-  "version": "7.19.23",
-  "description": "Read JSON files from api-extractor, generate documentation pages",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/microsoft/rushstack.git",
-    "directory": "apps/api-documenter"
-  },
-  "homepage": "https://api-extractor.com/",
-  "license": "MIT",
-  "scripts": {
-    "build": "heft build --clean",
-    "_phase:build": "heft build --clean",
-    "_phase:test": "heft test --no-build"
-  },
-  "bin": {
-    "api-documenter": "./bin/api-documenter"
-  },
-  "main": "lib/index.js",
-  "typings": "dist/rollup.d.ts",
+	"name": "typescript-reference-doc-generator",
+	"version": "0.0.1",
+	"description": "Reads JSON files from api-extractor, generate documentation pages",
+	"author": "The WordPress Contributors",
+	"license": "Apache-2.0",
+	"keywords": [
+		"wordpress",
+		"php",
+		"wasm",
+		"WebAssembly"
+	],
+	"homepage": "https://github.com/WordPress/wordpress-wasm/tree/HEAD/packages/wordpress-wasm",
+	"repository": {
+		"type": "git",
+		"url": "https://github.com/WordPress/wordpress-wasm.git",
+		"directory": "packages/typescript-reference-doc-generator"
+	},
+	"bugs": {
+		"url": "https://github.com/WordPress/wordpress-wasm/issues"
+	},
+	"engines": {
+		"node": ">=12"
+	},
+	"main": "build/index.js",
+	"types": "build-types",
+	"publishConfig": {
+		"access": "public"
+	},
   "dependencies": {
     "@microsoft/api-extractor-model": "7.25.2",
     "@microsoft/tsdoc": "0.14.2",
-    "@rushstack/node-core-library": "3.53.2",
-    "@rushstack/ts-command-line": "4.13.0",
     "colors": "~1.2.1",
-    "js-yaml": "~3.13.1",
-    "resolve": "~1.17.0",
     "typescript": "4.8.4"
   },
   "devDependencies": {
     "@rushstack/eslint-config": "3.1.1",
-    "@types/node": "12.20.24",
-    "@types/resolve": "1.17.1"
+    "@types/node": "12.20.24"
   }
 }
