+++ packages/typescript-reference-doc-generator/src/documenters/DocumenterConfig.ts
@@ -2,11 +2,7 @@
 // See LICENSE in the project root for license information.
 
 import * as path from 'path';
-import {
-	JsonSchema,
-	JsonFile,
-	NewlineKind,
-} from '@rushstack/node-core-library';
+import * as fs from 'fs';
 import { IConfigFile } from './IConfigFile';
 
 /**
@@ -19,19 +15,6 @@ export class DocumenterConfig {
 	public readonly configFile: IConfigFile;
 
 	/**
-	 * Specifies what type of newlines API Documenter should use when writing output files.  By default, the output files
-	 * will be written with Windows-style newlines.
-	 */
-	public readonly newlineKind: NewlineKind;
-
-	/**
-	 * The JSON Schema for API Documenter config file (api-documenter.schema.json).
-	 */
-	public static readonly jsonSchema: JsonSchema = JsonSchema.fromFile(
-		path.join(__dirname, '..', 'schemas', 'api-documenter.schema.json')
-	);
-
-	/**
 	 * The config file name "api-documenter.json".
 	 */
 	public static readonly FILENAME: string = 'api-documenter.json';
@@ -39,18 +22,6 @@ export class DocumenterConfig {
 	private constructor(filePath: string, configFile: IConfigFile) {
 		this.configFilePath = filePath;
 		this.configFile = configFile;
-
-		switch (configFile.newlineKind) {
-			case 'lf':
-				this.newlineKind = NewlineKind.Lf;
-				break;
-			case 'os':
-				this.newlineKind = NewlineKind.OsDefault;
-				break;
-			default:
-				this.newlineKind = NewlineKind.CrLf;
-				break;
-		}
 	}
 
 	/**
@@ -59,9 +30,8 @@ export class DocumenterConfig {
 	 * @param  configFilePath
 	 */
 	public static loadFile(configFilePath: string): DocumenterConfig {
-		const configFile: IConfigFile = JsonFile.loadAndValidate(
-			configFilePath,
-			DocumenterConfig.jsonSchema
+		const configFile: IConfigFile = JSON.parse(
+			fs.readFileSync(configFilePath).toString()
 		);
 
 		return new DocumenterConfig(path.resolve(configFilePath), configFile);
