+++ packages/typescript-reference-doc-generator/tsconfig.json
@@ -2,7 +2,6 @@
   "extends": "./node_modules/@rushstack/heft-node-rig/profiles/default/tsconfig-base.json",
 
   "compilerOptions": {
-    "types": ["heft-jest"],
     "lib": [
       "esnext"
     ]
